% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_data}
\alias{sf_data}
\title{Download a Sioux Falls dataset as CSV}
\usage{
sf_data(key, layer = NULL)
}
\arguments{
\item{key}{Dataset key from sf_datasets(), or a raw item_id}

\item{layer}{Layer number (only needed if using raw item_id)}
}
\value{
tibble with dataset records
}
\description{
Download a Sioux Falls dataset as CSV
}
