% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_search}
\alias{sf_search}
\title{Search Sioux Falls DataWorks portal}
\usage{
sf_search(q = "")
}
\arguments{
\item{q}{Search query}
}
\value{
tibble: id, type, title, description, url
}
\description{
Search Sioux Falls DataWorks portal
}
