% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_violent_crimes}
\alias{sf_violent_crimes}
\title{Get Sioux Falls violent crime statistics by year}
\usage{
sf_violent_crimes()
}
\value{
tibble with annual violent crime data
}
\description{
Get Sioux Falls violent crime statistics by year
}
